raise Exception

+++ cactus/demograficos/schemas/telefonoSchema.py
@@ -576,23 +576,19 @@ class SendSmsPin(graphene.Mutation):
 
     def mutate(self, info, telefono, registro_nuevo):
         if registro_nuevo:
-            try:
-                Telefono.objects.filter(
-                    telefono=telefono,
-                    validado=True
-                ).exclude(user=None).last().user
-                raise Exception("Este teléfono ya pertenece a un usuario")
-            except Exception:
-                Telefono.objects.filter(
-                    telefono=telefono,
-                    user=None).delete()
-                tel = Telefono.objects.create(
-                    telefono=telefono,
-                    activo=False,
-                    validado=False,
-                )
-                tel.send_token()
-                return SendSms(resp=True)
+            if User.objects.filter(username=telefono).count() > 0:
+                raise Exception("Telefono ya registrado en una cuenta \
+                    Inguz")
+            Telefono.objects.filter(
+                telefono=telefono,
+                user=None).delete()
+            tel = Telefono.objects.create(
+                telefono=telefono,
+                activo=False,
+                validado=False,
+            )
+            tel.send_token()
+            return SendSms(resp=True)
 
         else:
             try:
@@ -745,8 +741,11 @@ class ValidacionTelefono(graphene.Mutation):
         try:
             tel = Telefono.objects.get(telefono=numero)
         except Exception:
-            Exception('Número no registrado')
+            raise Exception('Número no registrado')
         if test or tel.is_valid(pin):
+            if enrolamiento:
+                if User.objects.filter(username=numero).count() > 0:
+                    raise Exception('Número con cuenta Inguz existente')
             if not enrolamiento:
                 try:
                     tel = Telefono.objects.filter(telefono=numero).exclude(
